feat: renamed nats package to more generic name - 'queue'
