+++ sub/internal/transport/queue/publisher/mailer/client.go
